主合同信息修改

+++ Ft-Cloud/ft-services/ft-service-pm/src/main/java/com/hhwy/pm/xmsl/contractInfo/service/IXmslContractInfoService.java
@@ -27,7 +27,7 @@ public interface IXmslContractInfoService {
      * @param xmslContractInfo
      * @return
      */
-    int insertXmslContractInfo(XmslContractInfo xmslContractInfo);
+    Long insertXmslContractInfo(XmslContractInfo xmslContractInfo);
 
     int insertXmslContractInfoList(List<XmslContractInfo> xmslContractInfoList);
